Make sure filesize is atleast the one declared by the browser
xhr.abort has some weird behavior on all browsers, if a XHR request
with a file upload is being aborted, the file gets sent anyway
even if it's only 30% uploaded. So the request is presumed
"completed" by the server and only the part of the file is
processed. This check tries to make sure that the full file
was received and that it was not an abort.